fix: Add products in ProductService🫸🌀✏️📗🐧🐳⬆

+++ common/src/main/java/id/my/hendisantika/common/Product.java
@@ -12,7 +12,7 @@ package id.my.hendisantika.common;
  * To change this template use File | Settings | File Templates.
  */
 public class Product {
-    private int id;
+    private long id;
     private String name;
     private String description;
     private double price;
@@ -20,18 +20,18 @@ public class Product {
     public Product() {
     }
 
-    public Product(int id, String name, String description, double price) {
+    public Product(long id, String name, String description, double price) {
         this.id = id;
         this.name = name;
         this.description = description;
         this.price = price;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
